Don't assume processes' used virtual address space starts at addr zero
Each process's 'struct proc' tracks the extent of its used virtual
address space. Previously this was a single 'sz' field and the used
space was assumed to start at address 0. Here we refactor that into
'vbase' and 'vlimit' fields. Currently, 'vbase' is always 0 and 'vlimit'
is equivalent to 'sz', but this change will make it possible to vary
that easily later.